+++ Checks/PWR040/example.f90
@@ -1,9 +1,10 @@
 ! PWR040: Consider loop tiling to improve the locality of reference
 
 pure subroutine example(a, b)
+  use iso_fortran_env, only: real32
   implicit none
-  real, dimension(:, :), intent(out) :: a
-  real, dimension(:, :), intent(in) :: b
+  real(kind=real32), dimension(:, :), intent(out) :: a
+  real(kind=real32), dimension(:, :), intent(in) :: b
   integer :: i, j
 
   do j = 1, size(a, 2)
