+++ Checks/PWR051/example-scalar.f90
@@ -2,9 +2,10 @@
 
 ! NOT-PWR003: `pure` prevents OpenMP parallelization
 function example(A) result(sum)
+  use iso_fortran_env, only: real32
   implicit none
-  real(kind=8), intent(in) :: A(:)
-  real(kind=8) :: sum
+  real(kind=real32), intent(in) :: A(:)
+  real(kind=real32) :: sum
   integer :: i
 
   sum = 0.0
