+++ Checks/PWR054/example-scalar.f90
@@ -1,9 +1,10 @@
 ! PWR054: consider applying vectorization to scalar reduction loop
 
 pure function example(A) result(sum)
+  use iso_fortran_env, only: real32
   implicit none
-  real(kind=8), intent(in) :: A(:)
-  real(kind=8) :: sum
+  real(kind=real32), intent(in) :: A(:)
+  real(kind=real32) :: sum
   integer :: i
 
   sum = 0.0
