+++ Checks/PWR060/example.f90
@@ -1,10 +1,11 @@
 ! PWR060 : consider loop fission to separate gather memory access pattern
 
 pure subroutine example(a, D, X, Y, index)
+  use iso_fortran_env, only: real32
   implicit none
-  real(kind=8), intent(in) :: a
-  real(kind=8), dimension(:), intent(out) :: D
-  real(kind=8), dimension(:), intent(in) :: X, Y
+  real(kind=real32), intent(in) :: a
+  real(kind=real32), dimension(:), intent(out) :: D
+  real(kind=real32), dimension(:), intent(in) :: X, Y
   integer, dimension(:), intent(in) :: index
   integer :: i
